backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.0
groq>=1.1.0
slowapi==0.1.9
pydantic-settings==2.6.0
python-dotenv==1.0.1
pytest==8.3.0
httpx==0.28.0
pytest-asyncio==0.24.0
